- Updated the code with main branch

+++ src/GetBoundingRectange.py
@@ -1,61 +1,72 @@
 import cv2
 import numpy as np
 import os
-import tqdm, csv
+import csv
 
 """
 A function to get the binary black and white image from a RGB image
 """
-def BGR2BINARY (image):
+
+
+def BGR2BINARY(image, x, y):
     # convert to gray scale
-    gray_image = cv2.cvtColor(image,cv2.COLOR_BGR2GRAY)
+    gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
     # apply the threshold
-    blur = cv2.GaussianBlur(gray_image,(5,5),0)
-    _, thresh_image = cv2.threshold(blur,0,255,cv2.THRESH_BINARY+cv2.THRESH_OTSU)
-    
-    # Negate the image to get a white background and black character
+    blur = cv2.GaussianBlur(gray_image, (5, 5), 0)
+    _, thresh_image = cv2.threshold(blur, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
+
+    # Negate the image to get a black background and white character
     binary_image = cv2.bitwise_not(thresh_image)
     # Apply opening to remove noise
-    kernel = np.ones((3,3),np.uint8)
+    kernel = np.ones((x, y), np.uint8)
     final_image = cv2.morphologyEx(binary_image, cv2.MORPH_OPEN, kernel)
     return final_image
+
 
 """
 A function to get the bounding rectange of the binary image
 """
+
+
 def getBoundingRect(image):
-    x1,y1,w,h = cv2.boundingRect(image)
-    
-    x2 = x1+w
-    y2 = y1+h
-    bounding_rect_image = image [y1:y2,x1:x2]
+    x1, y1, w, h = cv2.boundingRect(image)
+
+    x2 = x1 + w
+    y2 = y1 + h
+    bounding_rect_image = image[y1:y2, x1:x2]
+
+    # Return image to white background with
     bounding_rect_image = cv2.bitwise_not(bounding_rect_image)
     return bounding_rect_image
 
-def getContourRect(image):
-    contours, hir = cv2.findContours(image,1, cv2.CHAIN_APPROX_SIMPLE )
-    print('number of count: '+str(len(contours)))
-    if len(contours) > 1:
-        contours, _ = cv2.findContours(image,cv2.RETR_LIST,cv2.CHAIN_APPROX_SIMPLE)
+
+def getCountourRect(image):
+    countours, hir = cv2.findContours(image, 1, cv2.CHAIN_APPROX_SIMPLE)
+    print('number of count: ' + str(len(countours)))
+    if len(countours) > 1:
+        countours, _ = cv2.findContours(image, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
         h = 0
         img = 0
-        for c in contours:
+        for c in countours:
             x, y, w, h_temp = cv2.boundingRect(c)
             if h_temp > h:
                 h = h_temp
                 img = c
         x, y, w, h = cv2.boundingRect(img)
-        x2 = x+w
-        y2 = y+h
-        bounding_rect_image = image [y:y2,x:x2]
+        x2 = x + w
+        y2 = y + h
+        bounding_rect_image = image[y:y2, x:x2]
         bounding_rect_image = cv2.bitwise_not(bounding_rect_image)
         return bounding_rect_image
     else:
         return getBoundingRect(image)
 
+
 """
 A function to get the bounding box picture resized to a square
 """
+
+
 def resizeToSquare(boundingBox):
     box_height, box_width = boundingBox.shape
     if box_height >= box_width:
@@ -76,6 +87,8 @@ def resizeToSquare(boundingBox):
 """
 Method to resize square frame image
 """
+
+
 def resizeImage(img, height, width):
     return cv2.resize(img, (height, width))
 
@@ -83,17 +96,19 @@ def resizeImage(img, height, width):
 """
 Method that does all the previous steps in one function
 """
+
+
 def processImage(dir):
     n = dir.find('img')
     srcImg = cv2.imread(dir)
-    if dir[n+3:n+6] == '045' or dir[n+3:n+6] == '046':
+    if dir[n + 3:n + 6] == '045' or dir[n + 3:n + 6] == '046':
         print('i or j')
-        binaryImage = BGR2BINARY(srcImg)
+        binaryImage = BGR2BINARY(srcImg, 10, 10)
         boundingRect = getBoundingRect(binaryImage)
     else:
-        print('not i neither j') 
-        binaryImage = BGR2BINARY(srcImg)
-        boundingRect = getContourRect(binaryImage)
+        print('not i neither j')
+        binaryImage = BGR2BINARY(srcImg, 10, 10)
+        boundingRect = getCountourRect(binaryImage)
 
     return boundingRect
 
@@ -101,12 +116,15 @@ def processImage(dir):
 """
 Save images to directory
 """
+
+
 def saveImages(dirName):
     for filename in os.listdir(dirName):
         f = os.path.join('./' + dirName, filename)
         if os.path.isfile(f):
             toAdd = processImage(f, 30, 30)
             cv2.imwrite(f, toAdd)
+
 
 def getListOfCharacters():
     listOfCharacters = []
@@ -117,29 +135,34 @@ def getListOfCharacters():
 
     return listOfCharacters
 
+
 """
 Main function
 """
+
+
 def main():
-    dirname, filename = os.path.split(os.path.abspath(__file__))  
+    dirname, filename = os.path.split(os.path.abspath(__file__))
     pass
+
 
 if __name__ == '__main__':
     dirname, filename = os.path.split(os.path.abspath(__file__))
     list_of_Characters = getListOfCharacters()
 
     for idx, img in enumerate(list_of_Characters):
-        # print(idx)
+        print(idx)
         image_name = list_of_Characters[idx]['image']  # image name
         image_name = image_name[image_name.index('/') + 1:]
-        # print(image_name)
-        dir = dirname+f'\EnglishHandwrittenCharacters\{image_name}'
-        boundingRect = processImage(dir)
-        boundingRect = resizeToSquare(boundingRect)
-        boundingRect = cv2.resize(boundingRect, (30,30))
-        save_dir = dirname+f'\SquaredWithWhiteAdded\{image_name}'
-        cv2.imwrite(save_dir, boundingRect)
-    
+        if os.path.exists(dirname + f'\EnglishHandwrittenCharacters\{image_name}'):
+            # print(image_name)
+            dir = dirname + f'\EnglishHandwrittenCharacters\{image_name}'
+            boundingRect = processImage(dir)
+            # boundingRect = cv2.resize(boundingRect, (30,30))
+            boundingRect = resizeToSquare(boundingRect)
+            save_dir = dirname + f'\SquaredWithWhiteAdded\{image_name}'
+            cv2.imwrite(save_dir, boundingRect)
+
     # image_name = list_of_Characters[65]['image']
     # print(image_name)
     # image_name = image_name[image_name.index('/') + 1:]
@@ -151,10 +174,9 @@ if __name__ == '__main__':
     # cv2.imshow('BoundingRect',boundingRect)
     # cv2.waitKey(0)
     # cv2.destroyAllWindows()
-    # save_dir = dirname+f'\BoundingBoxes\{image_name}'
+    # save_dir = dirname+f'\SquaredWithWhiteAdded\{image_name}'
     # print(save_dir)
     # cv2.imwrite(save_dir, boundingRect)
-
 
     # boundingRect1 = cv2.resize(boundingRect, (30,30))
     # print(boundingRect[4])
@@ -174,4 +196,3 @@ if __name__ == '__main__':
     # cv2.imshow('frame',boundingRectangleImage)
     # cv2.waitKey(0)
     # print(boundingRectangleImage.shape)
-
